# public/openapi.yaml
openapi: "3.0.3"
info:
  title: Codepylot API
  version: "1.0.0"
  description: |
    Codepylot API — AI-powered sprint board that turns your ideas into shipped code.

    Authentication: Use `Authorization: Bearer <api-key>` header or session cookie.
servers:
  - url: /api
    description: API base
paths:
...
  /projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get project
      tags: [Projects]
      responses:
        "200":
          description: Project details
    patch:
      summary: Update project
      tags: [Projects]
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete project
      tags: [Projects]
      responses:
        "200":
          description: Deleted
  /projects/{projectId}/stories:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: List stories
      tags: [Stories]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [BACKLOG, TODO, IN_PROGRESS, REVIEW, DONE] }
        - name: sprintId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Array of stories
    post:
      summary: Create story
      tags: [Stories]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                priority: { type: string, enum: [CRITICAL, HIGH, MEDIUM, LOW] }
                status: { type: string }
                storyPoints: { type: integer }
                type: { type: string, enum: [feature, bug, chore, refactor, docs, test] }
      responses:
        "200":
          description: Created story
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []